refactor : Controller 클래스 패키지 이동, 클래스명 변경
Controller라는 클래스명 보다 사다리 게임 전체를 관리하는 의미로 LadderGame이 낫다고 판단해서 변경하고 패키지도 Controller라는 패키지를 따로 만드는 것보다 domain 패키지에 같이 있으면 되겠다고 생각했다.
